separated interfaces from abstract rpc class, changed id->key in KVS, objects to KVS[] instead of any[] in MPA_LISTING_ADD
